+++ src/main/java/com/fc/model/Message.java
@@ -10,12 +10,12 @@ public class Message {
     private Integer uid;
 
     //哪个用户
-    private Integer otherId;
+    private Long otherId;
     private String otherUsername;
     //什么操作
     private String operation;
     //操作了什么内容
-    private Integer postId;
+    private Long postId;
     //帖子，回复，评论
     private String displayedContent;
 
@@ -40,11 +40,11 @@ public class Message {
         this.uid = uid;
     }
 
-    public Integer getOtherId() {
+    public Long getOtherId() {
         return otherId;
     }
 
-    public void setOtherId(Integer otherId) {
+    public void setOtherId(Long otherId) {
         this.otherId = otherId;
     }
 
@@ -64,11 +64,11 @@ public class Message {
         this.operation = operation;
     }
 
-    public Integer getPostId() {
+    public Long getPostId() {
         return postId;
     }
 
-    public void setPostId(Integer postId) {
+    public void setPostId(Long postId) {
         this.postId = postId;
     }
